Update MapMenu runtime-data members
Members updated in this commit are used in MapMenu.as and LocalMap.as.
The BSTArray<GFxValue> is the content of WorldMap.MarkerData, and
changes depending on the function being called (CreateMarkers/RefreshMarkers).
The cameraPickDirection seems to be 3 floats used to calculate the custom
marker position based on cameraPickOrigin, updated when the custom
marker is created but only when the map is moved.