TESObjectWEAP research, based on the loader
Findings based on analysis of the TESObjectWEAP loader in LE and SSE. Several fields can be identified based on the fact that Skyrim supports both modern WEAP/DNAM (record versions >= 17) and FO3-format WEAP/DNAM.

In particular, `rumblePattern` was misidentified. The WEAP loader compares the values in RangedData to their defaults in order to know whether it should bother to allocate a RangedData at all; `rumblePattern` is one of those values and is compared as a float in both LE (via FPU opcodes) and SSE (via UCOMISS).

Additionally, the game computes how much condition damage a weapon would take (search for references to game settings fDamageToWeaponMeleeMult and fDamageToWeaponGunMult), and there, one can see the Flag2 that controls whether the "damage to weapon mult" field is used. (Not sure if the result of the condition damage calculation is ever actually applied anywhere. I suspect it isn't.)